Map, vessel location and types, password rules, codeless registration and notification decisions added

+++ src/lib/SMA/PAA/SERVICE/NotificationService.php
@@ -5,58 +5,158 @@ use SMA\PAA\InvalidParameterException;
 use SMA\PAA\ORM\NotificationModel;
 use SMA\PAA\ORM\NotificationRepository;
 use SMA\PAA\ORM\UserModel;
+use SMA\PAA\ORM\DecisionRepository;
 
 class NotificationService
 {
+    const MAX_NUM_OF_DECISIONS = 5;
+
+    private function validateDecision(string $portCallMasterId, int $imo, array $decisions)
+    {
+        $decisionRepository = new DecisionRepository();
+        $decisionRepository->validateSimplePortCallDecision($portCallMasterId, $imo);
+
+        if (count($decisions) > NotificationService::MAX_NUM_OF_DECISIONS) {
+            throw new InvalidParameterException(
+                "Too many decisions. Maximum is: " . NotificationService::MAX_NUM_OF_DECISIONS
+            );
+        }
+
+        foreach ($decisions as $decision) {
+            if (empty(trim($decision))) {
+                throw new InvalidParameterException(
+                    "Decision cannot be empty."
+                );
+            }
+        }
+    }
+
+    private function createDecision(
+        int $notificationId,
+        string $portCallMasterId,
+        array $decisionItemLabels
+    ) {
+        $decisionRepository = new DecisionRepository();
+        $decisionRepository->addSimplePortCallDecision($notificationId, $portCallMasterId, $decisionItemLabels);
+    }
+
     public function get(string $id)
     {
         $repository = new NotificationRepository();
         $model = $repository->get($id);
-        $model->sender = $this->getUser($model->created_by);
-        if ($model->ship_imo) {
-            $model->ship = $this->getVessel($model->ship_imo);
+
+        if ($model !== null) {
+            $model->sender = $this->getUser($model->created_by);
+            if ($model->ship_imo) {
+                $model->ship = $this->getVessel($model->ship_imo);
+            }
+            $model->decision = $model->getDecision();
+
+            $model->children = null;
+            $children = $model->getChildren();
+            foreach ($children as $child) {
+                $child->sender = $this->getUser($child->created_by);
+                $model->children[] = $child;
+            }
         }
-        if ($model->sender) {
-            unset($model->sender->last_login_time);
-        }
-        if ($model->sender) {
-            unset($model->sender->last_login_data);
-        }
-        if ($model->sender) {
-            unset($model->sender->last_session_time);
-        }
-        if ($model->sender) {
-            unset($model->sender->registration_code_id);
-        }
-        unset($model->created_by);
-        unset($model->modified_at);
-        unset($model->modified_by);
+
         return $model;
     }
 
-    public function add(string $type, string $message, string $ship_imo = null)
-    {
-        $validTypes = array("port", "ship");
+    public function add(
+        string $type,
+        string $message,
+        string $ship_imo = null,
+        string $port_call_master_id = null,
+        array $decisions = null,
+        int $parent_id = null
+    ) {
+        $validTypes = array(
+            NotificationModel::TYPE_PORT,
+            NotificationModel::TYPE_SHIP,
+            NotificationModel::TYPE_PORT_CALL_DECISION
+        );
+
         if (!$type) {
             throw new InvalidParameterException("Type can't be empty");
         } elseif (!in_array($type, $validTypes)) {
             throw new InvalidParameterException("Type should be one of there: " . implode(", ", $validTypes));
         }
+
         if (!$message) {
             throw new InvalidParameterException("Message can't be empty");
         }
-        if ($type === "port" && $ship_imo) {
-            throw new InvalidParameterException("Ship imo can't be assigned when message type is 'port'");
+
+        if (!empty($ship_imo)) {
+            if (!is_numeric($ship_imo)) {
+                throw new InvalidParameterException("Ship imo must be number");
+            }
         }
-        if ($type === "ship" && !is_numeric($ship_imo)) {
-            throw new InvalidParameterException("Ship imo must be number");
+
+        if ($type === NotificationModel::TYPE_PORT) {
+            if (!empty($ship_imo) ||
+                !empty($port_call_master_id) ||
+                !empty($decisions) ||
+                !empty($parent_id)) {
+                throw new InvalidParameterException(
+                    "Invalid assigned values for type: " . NotificationModel::TYPE_PORT
+                    . ". Valid assigned value is message."
+                );
+            }
+        } elseif ($type === NotificationModel::TYPE_SHIP) {
+            if (empty($ship_imo) || !empty($port_call_master_id) || !empty($decisions)) {
+                throw new InvalidParameterException(
+                    "Invalid assigned values for type: " . NotificationModel::TYPE_SHIP
+                    . ". Valid assigned values are message, ship_imo and parent_id."
+                );
+            }
+        } elseif ($type === NotificationModel::TYPE_PORT_CALL_DECISION) {
+            if (empty($ship_imo) ||
+                empty($port_call_master_id) ||
+                empty($decisions) ||
+                !empty($parent_id)) {
+                throw new InvalidParameterException(
+                    "Invalid assigned values for type: " . NotificationModel::TYPE_PORT_CALL_DECISION
+                    . ". Valid assigned values are message, ship_imo, port_call_master_id and decisions."
+                );
+            } else {
+                $this->validateDecision($port_call_master_id, $ship_imo, $decisions);
+            }
+        }
+
+        $repository = new NotificationRepository();
+
+        if (!empty($parent_id)) {
+            $parentModel = $repository->get($parent_id);
+
+            if ($parentModel === null) {
+                throw new InvalidParameterException(
+                    "Invalid parent ID: " . $parent_id
+                );
+            }
+
+            if ($parentModel->parent_notification_id !== null) {
+                throw new InvalidParameterException(
+                    "Parent ID: " . $parent_id . " already has parent. Only one level of hierarchy permitted."
+                );
+            }
+
+            if ($parentModel->type !== NotificationModel::TYPE_PORT_CALL_DECISION) {
+                throw new InvalidParameterException(
+                    "Parent ID: " . $parent_id . " has invalid type. Only " .
+                    NotificationModel::TYPE_PORT_CALL_DECISION . " allowed."
+                );
+            }
         }
 
         $id = null;
-        $repository = new NotificationRepository();
         $model = new NotificationModel();
-        $model->set($type, $message, $ship_imo);
+        $model->set($type, $message, $ship_imo, $parent_id);
         $id = $repository->save($model);
+
+        if ($type === NotificationModel::TYPE_PORT_CALL_DECISION) {
+            $this->createDecision($id, $port_call_master_id, $decisions);
+        }
 
         $notification = $this->get($id);
 
@@ -72,32 +172,44 @@ class NotificationService
     public function delete(int $id)
     {
         $repository = new NotificationRepository();
-        return $repository->delete([$id]);
+        $model = $repository->get($id);
+
+        if ($model === null) {
+            throw new InvalidParameterException(
+                "Cannot find notification with given id: " . $id
+            );
+        }
+
+        $repository->delete([$id]);
+
+        return ["result" => "OK"];
     }
 
     public function list()
     {
-        $repository = new NotificationRepository();
-        return array_map(function ($object) {
-            $object->sender = $this->getUser($object->created_by);
-            if ($object->ship_imo) {
-                $object->ship = $this->getVessel($object->ship_imo);
-            }
-            if ($object->sender) {
-                unset($object->sender->last_login_time);
-                unset($object->sender->last_login_data);
-                unset($object->sender->last_session_time);
-                unset($object->sender->registration_code_id);
+        $res = [];
 
-                if ($object->sender->status === UserModel::STATUS_DELETED) {
-                    $object->sender->email = "ex-user";
-                }
+        $repository = new NotificationRepository();
+        $query["parent_notification_id"] = null;
+        $rawResults = $repository->list($query, 0, 100, "created_at DESC");
+        foreach ($rawResults as $rawResult) {
+            $rawResult->sender = $this->getUser($rawResult->created_by);
+            if ($rawResult->ship_imo) {
+                $rawResult->ship = $this->getVessel($rawResult->ship_imo);
             }
-            unset($object->created_by);
-            unset($object->modified_at);
-            unset($object->modified_by);
-            return $object;
-        }, $repository->list([], 0, 100, "created_at DESC"));
+            $rawResult->decision = $rawResult->getDecision();
+
+            $rawResult->children = null;
+            $children = $rawResult->getChildren();
+            foreach ($children as $child) {
+                $child->sender = $this->getUser($child->created_by);
+                $rawResult->children[] = $child;
+            }
+
+            $res[] = $rawResult;
+        }
+
+        return $res;
     }
 
     private function getClassCached($key, $group, $callback)
@@ -119,7 +231,7 @@ class NotificationService
     {
         return $this->getClassCached("vessel", $imo, function () use ($imo) {
             $service = new VesselService();
-            return $service->vessel($imo);
+            return $service->getForNotification($imo);
         });
     }
 }
