+++ src/lib/SMA/PAA/SERVICE/StateService.php
@@ -2,6 +2,7 @@
 namespace SMA\PAA\SERVICE;
 
 use SMA\PAA\Session;
+use SMA\PAA\ORM\SeaChartFixedVesselRepository;
 
 class StateService implements IStateService
 {
@@ -14,10 +15,14 @@ class StateService implements IStateService
         }
         return null;
     }
-    public function set(string $key, $data)
+    public function set(string $key, $data, int $expires = null)
     {
         $client = new RedisClient();
-        return $client->set($key, serialize($data));
+        $out = $client->set($key, serialize($data));
+        if ($expires !== null) {
+            $client->expire($key, $expires);
+        }
+        return $out;
     }
     public function getSet(string $key, callable $callback, int $expires = null)
     {
@@ -47,13 +52,36 @@ class StateService implements IStateService
         $this->set(self::LATEST_PORT_CALL_IMOS, array_map(function ($data) {
             return $data["imo"];
         }, $datas));
+        $this->set(self::LATEST_PORT_CALL_DETAILS, $datas);
+    }
+    public function rebuildInitialSharedData()
+    {
+        $this->rebuildSeaChartFixedVesselsLists();
+        $this->rebuildActivePortCallLists();
+    }
+    public function rebuildSeaChartFixedVesselsLists()
+    {
+        $fixedVesselRepository = new SeaChartFixedVesselRepository();
+        $fixedVessels = $fixedVesselRepository->getFixedVessels();
+        $resultMappings = array();
+        foreach ($fixedVessels as $fixedVessel) {
+            if (!empty($fixedVessel->mmsi)) {
+                $resultMappings[$fixedVessel->mmsi] = $fixedVessel->imo;
+            }
+        }
+        $this->set(self::SEA_CHART_FIXED_VESSELS_MMSI_IMO_MAP, $resultMappings);
     }
     public function triggerPortCalls()
     {
-        $this->delete(self::LATEST_PORT_CALLS);
-        $this->rebuildActivePortCallLists();
-        $sse = new SseService();
-        $sse->trigger("portcalls", "changed", []);
+        if ($this->get(self::LATEST_PORT_CALLS_LOCKED) === null) {
+            $this->set(self::LATEST_PORT_CALLS_LOCKED, true, 60);
+            $this->delete(self::LATEST_PORT_CALLS);
+            $this->rebuildActivePortCallLists();
+            $sse = new SseService();
+            $sse->trigger("portcalls", "changed", []);
+            $sse->trigger("queue-portcalls", "changed", []);
+            $this->delete(self::LATEST_PORT_CALLS_LOCKED);
+        }
     }
     public function triggerLogistics()
     {
